requirements.txt
asgiref==3.7.2
certifi==2024.2.2
distlib==0.3.8
Django==5.0.2
django-auth-token==0.1.9
django-chamber==0.6.15
django-choice-enumfields==1.1.3
django-cors-headers==4.3.1
django-generic-m2m-field==0.0.5
django-ipware==6.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drfpasswordless==1.5.9
filelock==3.13.1
filemagic==1.6
import-string==0.1.0
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2==2.9.9
PyJWT==2.8.0
PyPrind==2.11.3
python-ipware==2.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.10.0
Unidecode==1.3.8
virtualenv==20.25.0
